requirements.txt
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.2
django-disqus==0.5
django-paypal==0.3.6
django-toolbelt==0.0.1
django-widget-tweaks==1.4.1
gunicorn==19.7.1
idna==2.5
mock==2.0.0
olefile==0.44
pbr==3.1.1
Pillow==4.1.1
psycopg2==2.7.1
python-decouple==3.0
pytz==2017.2
requests==2.18.1
six==1.10.0
static3==0.7.0
stripe==1.62.0
urllib3==1.21.1
whitenoise==3.3.0
